# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "DeleteInput": {
        "properties": {
          "confirm": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "previewHash": {
            "type": "string"
          }
        },
        "required": [
          "previewHash",
          "password",
          "confirm"
        ],
        "type": "object"
      },
...
      "MemberInput": {
        "properties": {
          "password": {
            "type": "string"
          },
          "roles": {
            "items": {
              "type": "string"
            },
            "type": [
              "null",
              "array"
            ]
          },
          "userName": {
            "type": "string"
          }
        },
        "required": [
          "userName",
          "password"
        ],
        "type": "object"
      },
      "MemberRolesInput": {
        "properties": {
          "reason": {
            "type": "string"
          },
          "roles": {
            "items": {
              "type": "string"
            },
            "type": "array"
          }
        },
        "required": [
          "roles",
          "reason"
        ],
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Learning.Api | v1",
    "version": "1.0.0"
  },
  "openapi": "3.1.1",
  "paths": {
...
    "/api/v1/family/members": {
      "get": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "properties": {
                      "accountId": {
                        "format": "uuid",
                        "type": "string"
                      },
                      "id": {
                        "format": "uuid",
                        "type": "string"
                      },
                      "roles": {
                        "type": [
                          "null",
                          "string"
                        ]
                      },
                      "userName": {
                        "type": [
                          "null",
                          "string"
                        ]
                      }
                    },
                    "required": [
                      "id",
                      "accountId",
                      "roles",
                      "userName"
                    ],
                    "type": "object"
                  },
                  "type": "array"
                }
              }
            }
          }
        },
        "tags": [
          "Learning.Api"
        ]
      },
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MemberInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "accountId": {
                      "format": "uuid",
                      "type": "string"
                    },
                    "id": {
                      "format": "uuid",
                      "type": "string"
                    },
                    "roles": {
                      "type": [
                        "null",
                        "string"
                      ]
                    },
                    "userName": {
                      "type": [
                        "null",
                        "string"
                      ]
                    }
                  },
                  "required": [
                    "id",
                    "accountId",
                    "roles",
                    "userName"
                  ],
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "Learning.Api"
        ]
      }
    },
    "/api/v1/family/members/{id}/roles": {
      "put": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MemberRolesInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "id": {
                      "format": "uuid",
                      "type": "string"
                    },
                    "notice": {
                      "type": [
                        "null",
                        "string"
                      ]
                    },
                    "roles": {
                      "type": [
                        "null",
                        "string"
                      ]
                    }
                  },
                  "required": [
                    "id",
                    "roles",
                    "notice"
                  ],
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "Learning.Api"
        ]
      }
    },
    "/api/v1/family:delete": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "backupPolicy": {
                      "type": [
                        "null",
                        "string"
                      ]
                    },
                    "deleted": {
                      "type": "boolean"
                    },
                    "familyId": {
                      "format": "uuid",
                      "type": "string"
                    },
                    "receiptId": {
                      "format": "uuid",
                      "type": "string"
                    },
                    "scope": {
                      "type": [
                        "null",
                        "string"
                      ]
                    }
                  },
                  "required": [
                    "deleted",
                    "familyId",
                    "receiptId",
                    "scope",
                    "backupPolicy"
                  ],
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "Learning.Api"
        ]
      }
    },
...
  },
...
}